backend/requirements.txt
fastapi==0.115.2
uvicorn==0.30.1
python-multipart==0.0.9

# Google ADK + Gemini
google-generativeai==0.7.0
google-ai-generativelanguage==0.6.5

# Vector DB (Chroma)
chromadb==0.5.4

# Embeddings
sentence-transformers==3.0.1

# Supabase client
supabase==2.4.3

# Image / Video / PDF
pillow==10.3.0
opencv-python==4.9.0.80
numpy==1.26.4
pypdf==4.2.0

# Utils
python-dotenv==1.0.1
pydantic==2.7.3
pydantic-settings==2.1.0
loguru==0.7.2
httpx==0.27.0

# Testing
pytest==8.3.1
pytest-asyncio==0.23.6
